Flask/requirements.txt
flask
mysql-connector
celery
pymysql